pyproject.toml
[project]
name = "oauth-mcp-server"
version = "0.1.0"
description = "MCP Server acting as OAuth 2.1 Protected Resource"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.8.0",
    "starlette>=0.40.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "PyJWT[crypto]>=2.10.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "respx>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
